requirements.txt
# Core testing framework
pytest>=7.4.0
pytest-xdist>=3.3.0
pytest-reportportal>=5.3.0

# Playwright
playwright>=1.40.0

# Database connectivity (if using JDBC)
jpype1>=1.4.0

# Reporting and utilities
requests>=2.31.0

